+++ src/main/java/com/example/swip/repository/StudyRepository.java
@@ -1,7 +1,9 @@
 package com.example.swip.repository;
 
 import com.example.swip.entity.Study;
+import com.example.swip.repository.custom.StudyFilterRepositoryCustom;
+import com.example.swip.repository.custom.StudyRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudyRepository extends JpaRepository<Study, Long>, StudyFilterRepository, StudyRepositoryCustom{
+public interface StudyRepository extends JpaRepository<Study, Long>, StudyFilterRepositoryCustom, StudyRepositoryCustom {
 }
