+++ src/main/java/com/example/swip/repository/UserStudyRepository.java
@@ -2,9 +2,8 @@ package com.example.swip.repository;
 
 import com.example.swip.entity.UserStudy;
 import com.example.swip.entity.compositeKey.UserStudyId;
+import com.example.swip.repository.custom.UserStudyRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface UserStudyRepository extends JpaRepository<UserStudy, UserStudyId> , UserStudyRepositoryCustom{
+public interface UserStudyRepository extends JpaRepository<UserStudy, UserStudyId> , UserStudyRepositoryCustom {
 }
