+++ src/main/java/com/example/swip/repository/custom/StudyFilterRepositoryCustom.java
@@ -1,17 +1,15 @@
-package com.example.swip.repository;
+package com.example.swip.repository.custom;
 
 import com.example.swip.dto.quick_match.QuickMatchFilter;
 import com.example.swip.dto.quick_match.QuickMatchResponse;
 import com.example.swip.dto.study.StudyFilterCondition;
 import com.example.swip.dto.study.StudyFilterResponse;
 import com.example.swip.entity.Study;
-import com.querydsl.core.Tuple;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
-public interface StudyFilterRepository {
+public interface StudyFilterRepositoryCustom {
     List<StudyFilterResponse> filterStudy(StudyFilterCondition filterCondition);
     List<QuickMatchResponse> quickFilterStudy(QuickMatchFilter quickMatchFilter, Long userId, Long page, Long size);
     List<Study> progressStartStudy(LocalDate date);
