+++ src/main/java/com/example/swip/repository/custom/StudyTodoRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.swip.repository;
+package com.example.swip.repository.custom;
 
 import com.example.swip.entity.*;
 import com.querydsl.core.Tuple;
