+++ src/main/java/com/example/swip/repository/custom/UserStudyRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.swip.repository;
+package com.example.swip.repository.custom;
 
 import com.example.swip.entity.Study;
 import com.example.swip.entity.UserStudy;
